🧪 [testing improvement] Improve PdfPropertyGetter tests coverage
🎯 What: The PdfPropertyGetter had no dedicated test class and its test files were checked in as raw binaries causing repository bloat. Tests were scattered in PropertyGetterTests.cs.
📊 Coverage: Created PdfPropertyGetterTests to cover valid PDFs, empty PDFs, encrypted PDFs, missing files, and invalid formats. Implemented PdfSetup using Base64 encoded strings to dynamically create test files.
✨ Result: Enhanced test maintainability and portability, avoided raw binaries in source control, and aligned PDF property getter testing with standard project patterns.
