+++ HDLG.Tests/PdfSetup.cs
@@ -1,0 +1,38 @@
+using System;
+using System.IO;
+
+namespace HDLG.Tests
+{
+    internal static class PdfSetup
+    {
+        private static bool _initialized = false;
+        private static readonly object _lock = new();
+
+        public static void CreatePdfDocs()
+        {
+            if (_initialized) return;
+            lock (_lock)
+            {
+                if (_initialized) return;
+
+                string validPdfBase64 = "JVBERi0xLjMKJVx4OTNceDhjXHg4Ylx4OWUgUmVwb3J0TGFiIEdlbmVyYXRlZCBQREYgZG9jdW1lbnQgKG9wZW5zb3VyY2UpCjEgMCBvYmoKPDwKL0YxIDIgMCBSCj4+CmVuZG9iagoyIDAgb2JqCjw8Ci9CYXNlRm9udCAvSGVsdmV0aWNhIC9FbmNvZGluZyAvV2luQW5zaUVuY29kaW5nIC9OYW1lIC9GMSAvU3VidHlwZSAvVHlwZTEgL1R5cGUgL0ZvbnQKPj4KZW5kb2JqCjMgMCBvYmoKPDwKL0NvbnRlbnRzIDcgMCBSIC9NZWRpYUJveCBbIDAgMCA1OTUuMjc1NiA4NDEuODg5OCBdIC9QYXJlbnQgNiAwIFIgL1Jlc291cmNlcyA8PAovRm9udCAxIDAgUiAvUHJvY1NldCBbIC9QREYgL1RleHQgL0ltYWdlQiAvSW1hZ2VDIC9JbWFnZUkgXQo+PiAvUm90YXRlIDAgL1RyYW5zIDw8Cgo+PgogIC9UeXBlIC9QYWdlCj4+CmVuZG9iago0IDAgb2JqCjw8Ci9QYWdlTW9kZSAvVXNlTm9uZSAvUGFnZXMgNiAwIFIgL1R5cGUgL0NhdGFsb2cKPj4KZW5kb2JqCjUgMCBvYmoKPDwKL0F1dGhvciAoYW5vbnltb3VzKSAvQ3JlYXRpb25EYXRlIChEOjIwMjYwNTI1MTg0NDQ4KzAwJzAwJykgL0NyZWF0b3IgKGFub255bW91cykgL0tleXdvcmRzICgpIC9Nb2REYXRlIChEOjIwMjYwNTI1MTg0NDQ4KzAwJzAwJykgL1Byb2R1Y2VyIChSZXBvcnRMYWIgUERGIExpYnJhcnkgLSBcKG9wZW5zb3VyY2VcKSkKICAvU3ViamVjdCAodW5zcGVjaWZpZWQpIC9UaXRsZSAoVGVzdCBQREYgVGl0bGUpIC9UcmFwcGVkIC9GYWxzZQo+PgplbmRvYmoKNiAwIG9iago8PAovQ291bnQgMSAvS2lkcyBbIDMgMCBSIF0gL1R5cGUgL1BhZ2VzCj4+CmVuZG9iago3IDAgb2JqCjw8Ci9GaWx0ZXIgWyAvQVNDSUk4NURlY29kZSAvRmxhdGVEZWNvZGUgXSAvTGVuZ3RoIDEwMAo+PgpzdHJlYW0KR2FwUWgwRT1GLDBVXEgzVFxwTllUXlFLaz90Yz5JUCw7VyNVMV4yM2loUEVNXz9DVzRLSVNpOTBFQy1wPlFrUnRlPTwlVkwlLCM3Uy5uXE9PPVoocy9XYUVGL2NaY0QnLjN+PmVuZHN0cmVhbQplbmRvYmoKeHJlZgowIDgKMDAwMDAwMDAwMCA2NTUzNSBmCjAwMDAwMDAwNjEgMDAwMDAgbgowMDAwMDAwMDkyIDAwMDAwIG4KMDAwMDAwMDE5OSAwMDAwMCBuCjAwMDAwMDA0MDIgMDAwMDAgbgowMDAwMDAwNDcwIDAwMDAwIG4KMDAwMDAwMDczNyAwMDAwMCBuCjAwMDAwMDA3OTYgMDAwMDAgbgp0cmFpbGVyCjw8Ci9JRApbPGQzM2UxZTc5MzVjNTViMjkzMDJhNTA2M2FlYWNkZjM4PjxkMzNlMWU3OTM1YzU1YjI5MzAyYTUwNjNhZWFjZGYzOD5dCiUgUmVwb3J0TGFiIGdlbmVyYXRlZCBQREYgZG9jdW1lbnQgLS0gZGlnZXN0IChvcGVuc291cmNlKQoKL0luZm8gNSAwIFIKL1Jvb3QgNCAwIFIKL1NpemUgOAo+PgpzdGFydHhyZWYKOTg2CiUlRU9GCg==";
+                string emptyPdfBase64 = "JVBERi0xLjMKJVx4OTNceDhjXHg4Ylx4OWUgUmVwb3J0TGFiIEdlbmVyYXRlZCBQREYgZG9jdW1lbnQgKG9wZW5zb3VyY2UpCjEgMCBvYmoKPDwKL0YxIDIgMCBSCj4+CmVuZG9iagoyIDAgb2JqCjw8Ci9CYXNlRm9udCAvSGVsdmV0aWNhIC9FbmNvZGluZyAvV2luQW5zaUVuY29kaW5nIC9OYW1lIC9GMSAvU3VidHlwZSAvVHlwZTEgL1R5cGUgL0ZvbnQKPj4KZW5kb2JqCjMgMCBvYmoKPDwKL0NvbnRlbnRzIDcgMCBSIC9NZWRpYUJveCBbIDAgMCA1OTUuMjc1NiA4NDEuODg5OCBdIC9QYXJlbnQgNiAwIFIgL1Jlc291cmNlcyA8PAovRm9udCAxIDAgUiAvUHJvY1NldCBbIC9QREYgL1RleHQgL0ltYWdlQiAvSW1hZ2VDIC9JbWFnZUkgXQo+PiAvUm90YXRlIDAgL1RyYW5zIDw8Cgo+PgogIC9UeXBlIC9QYWdlCj4+CmVuZG9iago0IDAgb2JqCjw8Ci9QYWdlTW9kZSAvVXNlTm9uZSAvUGFnZXMgNiAwIFIgL1R5cGUgL0NhdGFsb2cKPj4KZW5kb2JqCjUgMCBvYmoKPDwKL0F1dGhvciAoYW5vbnltb3VzKSAvQ3JlYXRpb25EYXRlIChEOjIwMjYwNTI1MTkxODI5KzAwJzAwJykgL0NyZWF0b3IgKGFub255bW91cykgL0tleXdvcmRzICgpIC9Nb2REYXRlIChEOjIwMjYwNTI1MTkxODI5KzAwJzAwJykgL1Byb2R1Y2VyIChSZXBvcnRMYWIgUERGIExpYnJhcnkgLSBcKG9wZW5zb3VyY2VcKSkKICAvU3ViamVjdCAodW5zcGVjaWZpZWQpIC9UaXRsZSAoKSAvVHJhcHBlZCAvRmFsc2UKPj4KZW5kb2JqCjYgMCBvYmoKPDwKL0NvdW50IDEgL0tpZHMgWyAzIDAgUiBdIC9UeXBlIC9QYWdlcwo+PgplbmRvYmoKNyAwIG9iago8PAovRmlsdGVyIFsgL0FTQ0lJODVEZWNvZGUgL0ZsYXRlRGVjb2RlIF0gL0xlbmd0aCAxMDAKPj4Kc3RyZWFtCkdhcFFoMEU9RiwwVVxIM1RccE5ZVF5RS2s/dGM+SVAsO1cjVTFeMjNpaFBFTV8/Q1c0S0lTaTkwRUMtcD5Ra1J0ZT08JVZMJSwjN1MublxPTz1aKHMvV2FFRi9jWmNEJy4zfj5lbmRzdHJlYW0KZW5kb2JqCnhyZWYKMCA4CjAwMDAwMDAwMDAgNjU1MzUgZgowMDAwMDAwMDYxIDAwMDAwIG4KMDAwMDAwMDA5MiAwMDAwMCBuCjAwMDAwMDAxOTkgMDAwMDAgbgowMDAwMDAwNDAyIDAwMDAwIG4KMDAwMDAwMDQ3MCAwMDAwMCBuCjAwMDAwMDA3MjMgMDAwMDAgbgowMDAwMDAwNzgyIDAwMDAwIG4KdHJhaWxlcgo8PAovSUQKWzxjNDdhZTA5MzQ4OTYzZGNjYWQzZTQ0YWRjZDgzMDUyOD48YzQ3YWUwOTM0ODk2M2RjY2FkM2U0NGFkY2Q4MzA1Mjg+XQolIFJlcG9ydExhYiBnZW5lcmF0ZWQgUERGIGRvY3VtZW50IC0tIGRpZ2VzdCAob3BlbnNvdXJjZSkKCi9JbmZvIDUgMCBSCi9Sb290IDQgMCBSCi9TaXplIDgKPj4Kc3RhcnR4cmVmCjk3MgolJUVPRgo=";
+                string encryptedPdfBase64 = "JVBERi0xLjcKJb/3ov4KMSAwIG9iago8PCAvRXh0ZW5zaW9ucyA8PCAvQURCRSA8PCAvQmFzZVZlcnNpb24gLzEuNyAvRXh0ZW5zaW9uTGV2ZWwgOCA+PiA+PiAvUGFnZU1vZGUgL1VzZU5vbmUgL1BhZ2VzIDMgMCBSIC9UeXBlIC9DYXRhbG9nID4+CmVuZG9iagoyIDAgb2JqCjw8IC9BdXRob3IgPGFmMzAyZmFiY2U0Zjg1NDRjM2RjOWViMjY3N2NjZjI4OTQ4YzYzOWI2ODBjYjBmYjFkZjQ4ZDNmMzY0ZmIyYjY+IC9DcmVhdGlvbkRhdGUgPDY2ZWNkYWJlYjM1ZTUwNDBjYjJlMzkxZDYwYjc4ZDBiZDVhMTRiNzJjNWFiMWIyYTJhMTkwMGYxZGJhN2RjODZiZTBiMzQ1ZWQ0NGI0ZTljNjVlYzA1NmM0MTQyZWU1OD4gL0NyZWF0b3IgPDQxMDcxYjkxNmUwMjQ5NTRjYjZjNjRhMzAyM2YxOTgzNTVjMGFhYmI5NzA5MzdmMWZlNTJkNzYzZTdhNzRhMDk+IC9LZXl3b3JkcyA8YjYwYzdmMTEwYTRlMWEzOWIyNDM5YmZiODcyM2JiMmM4ZDFkMTkyNjk0OGExZGQwNjBiYmJlMDgxY2QyOThiNj4gL01vZERhdGUgPDNkNjA3MzQ4ZGIyNzc1NDFiZDJkNGVkMTk5ZmZkOWVlZjY4OTEwNzA3ODI3OGM2Mzg4ZWNlZDc1OTQwOTg4MmI0YjQxNGUzNzZmOWQ2M2QwNTE0NjM3OGYxZGQ2NzlkNz4gL1Byb2R1Y2VyIDw2YzI5NmY5MmNmNDA0ZjZkMjYwNjY0ZDlkZWIwMmE2YmI5NzFiY2ExN2U5NzUzZDkzNWVkZmFmYTI2NGQ1ZWQ5ZTZkNGIzODViZTY0Njg1N2Y4N2QzYmI3MGQ2NWI1MGQyMWI1MWIzZTgzNWIwNmYzYTRjMDYwZjc1NTY2ZmIzYj4gL1N1YmplY3QgPDA2ZmJhNTBhMDM3MDJjZGNhYzQwOGY4NWQ0ZWNkNzBiZDRhZWRhNjU5OGMwYjBhYmFjNTcyZjU5M2JhZTlhZmY+IC9UaXRsZSA8ZmQyMDk5NThhMTE1YjhkMzUxNWJkMGYwNzMxNjMzYmQwMmU2ZTM3ZWQ1NGU1NmQ5ODM3Y2E0ODUwM2FmZWExMD4gL1RyYXBwZWQgL0ZhbHNlID4+CmVuZG9iagozIDAgb2JqCjw8IC9Db3VudCAxIC9LaWRzIFsgNCAwIFIgXSAvVHlwZSAvUGFnZXMgPj4KZW5kb2JqCjQgMCBvYmoKPDwgL0NvbnRlbnRzIDUgMCBSIC9NZWRpYUJveCBbIDAgMCA1OTUuMjc1NiA4NDEuODg5OCBdIC9QYXJlbnQgMyAwIFIgL1Jlc291cmNlcyA8PCAvRm9udCA2IDAgUiAvUHJvY1NldCBbIC9QREYgL1RleHQgL0ltYWdlQiAvSW1hZ2VDIC9JbWFnZUkgXSA+PiAvUm90YXRlIDAgL1RyYW5zIDw8ID4+IC9UeXBlIC9QYWdlID4+CmVuZG9iago1IDAgb2JqCjw8IC9MZW5ndGggOTYgL0ZpbHRlciAvRmxhdGVEZWNvZGUgPj4Kc3RyZWFtCou2UY8LWAbFQGa23F0mKDhxeuu2GXO5jQjDGdIPKv34fY/mS5bz84tyv6YDau1wK+qxzzdr3yLIrTGXFACaOgJx4Lzok1q6/3u+0VUChSVIHuj8bQYVr1RmGuq1A3CCxmVuZHN0cmVhbQplbmRvYmoKNiAwIG9iago8PCAvRjEgNyAwIFIgPj4KZW5kb2JqCjcgMCBvYmoKPDwgL0Jhc2VGb250IC9IZWx2ZXRpY2EgL0VuY29kaW5nIC9XaW5BbnNpRW5jb2RpbmcgL05hbWUgL0YxIC9TdWJ0eXBlIC9UeXBlMSAvVHlwZSAvRm9udCA+PgplbmRvYmoKOCAwIG9iago8PCAvQ0YgPDwgL1N0ZENGIDw8IC9BdXRoRXZlbnQgL0RvY09wZW4gL0NGTSAvQUVTVjMgL0xlbmd0aCAzMiA+PiA+PiAvRmlsdGVyIC9TdGFuZGFyZCAvTGVuZ3RoIDI1NiAvTyA8YjgxNDUxMzRlNDQyYzMzMGNhNTQyNDNkOTRjOGE5MTVkYjM2ODdlNWE1OTRkMmQ2YjRjZTM2NzZkZDIwMDhiNTk3MTc4MTU2MGYwMGE5MWRjYmE0YzJmYmNjZDlhZDE4PiAvT0UgPDRlY2JhMzE1YTdlM2QyMzY1ZjMzODAyNDVjOTA0MTkwYmE0MjcwZWM5ZGJiOGZmMjI0MWRjZmY4MWRlYjkxMTQ+IC9QIC00IC9QZXJtcyA8YjFjNWJiY2Y5NDEyOGE5ZDViZDJkYjBmMTczM2MzMjg+IC9SIDYgL1N0bUYgL1N0ZENGIC9TdHJGIC9TdGRDRiAvVSA8NGY4ZTA1MmJmZTNiYWYzMjcwZDY4MzA4ZTRmZDdkN2Y5MzlmOTU1NzAxZDVmZmQ3NDJmOGJmMTUwMWVkYjcxYTQ3YTdhYWNkYzlhN2M1MjlkNzA1YmVkOWI4M2NhNmRjPiAvVUUgPDY1MjY5NjRkYmYwY2E0ZDEzZmQ4MmIyMGFlZmY2NWEwMzFlYjk3ODVlN2FhZmZlNzgxZDdiNmFlODA1YTA2MTU+IC9WIDUgPj4KZW5kb2JqCnhyZWYKMCA5CjAwMDAwMDAwMDAgNjU1MzUgZiAKMDAwMDAwMDAxNSAwMDAwMCBuIAowMDAwMDAwMTQ5IDAwMDAwIG4gCjAwMDAwMDA5MjYgMDAwMDAgbiAKMDAwMDAwMDk4NSAwMDAwMCBuIAowMDAwMDAxMTg0IDAwMDAwIG4gCjAwMDAwMDEzNTAgMDAwMDAgbiAKMDAwMDAwMTM4MSAwMDAwMCBuIAowMDAwMDAxNDg4IDAwMDAwIG4gCnRyYWlsZXIgPDwgL0luZm8gMiAwIFIgL1Jvb3QgMSAwIFIgL1NpemUgOSAvSUQgWzxkMzNlMWU3OTM1YzU1YjI5MzAyYTUwNjNhZWFjZGYzOD48OTkzNDY1NDA2NjYwYWZkMzY2N2E4M2IyZDM2YWI5Yzc+XSAvRW5jcnlwdCA4IDAgUiA+PgpzdGFydHhyZWYKMjAzNQolJUVPRgo=";
+                string invalidPdfBase64 = "aW52YWxpZCBkYXRh";
+
+                File.WriteAllBytes("test.pdf", Convert.FromBase64String(validPdfBase64));
+                File.WriteAllBytes("test_empty_title.pdf", Convert.FromBase64String(emptyPdfBase64));
+                File.WriteAllBytes("test_encrypted.pdf", Convert.FromBase64String(encryptedPdfBase64));
+                File.WriteAllBytes("test_invalid.pdf", Convert.FromBase64String(invalidPdfBase64));
+
+                _initialized = true;
+            }
+        }
+
+        public static void Cleanup()
+        {
+            // Do not delete files during parallel test execution to prevent race conditions.
+            // These temporary files will be cleaned up by the build/clean system.
+        }
+    }
+}
