--- HDLG.Tests/test_empty_title.pdf
@@ -1,68 +1,0 @@
-%PDF-1.3
-%\x93\x8c\x8b\x9e ReportLab Generated PDF document (opensource)
-1 0 obj
-<<
-/F1 2 0 R
->>
-endobj
-2 0 obj
-<<
-/BaseFont /Helvetica /Encoding /WinAnsiEncoding /Name /F1 /Subtype /Type1 /Type /Font
->>
-endobj
-3 0 obj
-<<
-/Contents 7 0 R /MediaBox [ 0 0 595.2756 841.8898 ] /Parent 6 0 R /Resources <<
-/Font 1 0 R /ProcSet [ /PDF /Text /ImageB /ImageC /ImageI ]
->> /Rotate 0 /Trans <<
-
->>
-  /Type /Page
->>
-endobj
-4 0 obj
-<<
-/PageMode /UseNone /Pages 6 0 R /Type /Catalog
->>
-endobj
-5 0 obj
-<<
-/Author (anonymous) /CreationDate (D:20260525191829+00'00') /Creator (anonymous) /Keywords () /ModDate (D:20260525191829+00'00') /Producer (ReportLab PDF Library - \(opensource\))
-  /Subject (unspecified) /Title () /Trapped /False
->>
-endobj
-6 0 obj
-<<
-/Count 1 /Kids [ 3 0 R ] /Type /Pages
->>
-endobj
-7 0 obj
-<<
-/Filter [ /ASCII85Decode /FlateDecode ] /Length 100
->>
-stream
-GapQh0E=F,0U\H3T\pNYT^QKk?tc>IP,;W#U1^23ihPEM_?CW4KISi90EC-p>QkRte=<%VL%,#7S.n\OO=Z(s/WaEF/cZcD'.3~>endstream
-endobj
-xref
-0 8
-0000000000 65535 f
-0000000061 00000 n
-0000000092 00000 n
-0000000199 00000 n
-0000000402 00000 n
-0000000470 00000 n
-0000000723 00000 n
-0000000782 00000 n
-trailer
-<<
-/ID
-[<c47ae09348963dccad3e44adcd830528><c47ae09348963dccad3e44adcd830528>]
-% ReportLab generated PDF document -- digest (opensource)
-
-/Info 5 0 R
-/Root 4 0 R
-/Size 8
->>
-startxref
-972
-%%EOF
