--- HDLG.Tests/test_invalid.pdf
@@ -1,1 +1,0 @@
-This is a fake pdf file.
